Update patch set 1
Patch Set 1:

> > (1 comment)
 > >
 > > not saying it's wrong to add it, but we probably need to choose
 > > what our effective minimum is
 > 
 > Because Centos7 defaults setuptools version 0.9.8 so I add the
 > limit for it.

The way it's added here is saying we cannot use the version in CentOS. Is that what we want? I assume we always want to use a newer version if we can. Is the goal to force an upgrade?

Patch-set: 1
Reviewer: Gerrit User 2472 <2472@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0