Add salt to global requirements
The heat-agents package uses salt as a test dependency, because it
contains an agent that can use SaltStack to do software deployments.
This previously did not cause requirements gate jobs to fail, but it now
does so add it to the global list.

Is the library actively maintained?
    Yes - last release < 3 months ago

Is the library good code?
    What is goodness?

Is the library python 3 compatible?
    Yes

Is the library license compatible?
    Yes - ASL2

Is the library already packaged in the distros we target (Ubuntu latest
/ Fedora latest)?

    No, and in practice it gets installed from pip by disk-image-builder
    anyway - there's no need for it to be co-installable with OpenStack
    since it's run on guest machines

Is the function of this library already covered by other libraries in
global-requirements.txt?

    No

Is the library required for OpenStack project or related dev or
infrastructure setup? (Answer to this should be Yes, of course) Which?

    Yes, for testing heat-agents

Change-Id: I5f37f512ba9168372e1085931b4784b8ec198ea0