Update patch set 2
Patch Set 2:

To be clear, it's heat-agents (i.e. a guest agent), not Heat itself.

I'm not sure I agree that the different agents need to be consistent. In the case of Salt there's an ASL2-licensed Python client library so the Salt agent uses it; in the case of Ansible there isn't so the Ansible agent uses the CLI.

Patch-set: 2
Label: Verified=0