Update patch set 2
Patch Set 2:

test-requirements.txt still has stuff with no minimum. test-requirements.txt also adds formencode.  Which is not in our global-requirements.txt (as all direct dependencies need to be).  That would be another review to add it, and I'm not sure if we already have a lib that provides the same functionality.

Those are my two issues, so sticking with the -1 for now.

Patch-set: 2