Add twine to global requirements
We have several jobs that run "python setup.py check -r -s" to validate
the repo's README file for issues that could cause PyPI upload failures.
This method is being deprecated with the new recommendation of running
"twine check" instead.

Is the library actively maintained?
    Yes - last release < 3 months ago, multiple releases per year

Is the library good code?
    Yes, maintained by the PyPa developers

Is the library python 3 compatible?
    Yes

Is the library license compatible?
    Yes - Apache 2 licensed

Is the library already packaged in the distros we target (Ubuntu latest
/ Fedora latest)?

    Yes, I see Ubuntu and Fedora packages

Is the function of this library already covered by other libraries in
global-requirements.txt?

    No

Is the library required for OpenStack project or related dev or
infrastructure setup? (Answer to this should be Yes, of course) Which?

    Yes, for package description validation

Change-Id: Ifd77e1218b47fad42880377802336df7285274ac
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>