Add tornado to global requirements
The monasca-analytics package tornado uses as a web service.
Add tornado to global list because requirements gate jobs to fail.
tornado was already listed in upper-constraints.txt.

Is the library actively maintained?
    Yes, last release about 2 months ago

Is the library good code?
    Yes

Is the library python 3 compatible?
    Yes

Is the library license compatible?
    Yes - Apache License 2.0

Is the library already packaged in the distros we target (Ubuntu latest
/ Fedora latest)?
    Yes, There is a pypi package available.

Is the function of this library already covered by other libraries in
global-requirements.txt?
    No

Is the library required for OpenStack project or related dev or
infrastructure setup?
    Yes, for monasca-analytics

If the library release is managed by the Openstack release process
does it use the cycle-with-intermediary release type?
    Yes

Do I need to update anything else?
    No

Change-Id: I21bed8f990caf4dc9fbee7b1ab602f6552ba83bd