Update patch set 2
Patch Set 2:

> I think the best option here would be to request requests to tag
 > and move to that.  rolling back has caused problems in the past
 > that I'd like to avoid.  This is what we've done in the past
 > (unless the release would take a while).

Actually it's what we've done sometimes but not in scenarios like this.  That satement is factually wrong.  The response to "this version of a library is broken for some way" shouldn't be "try a new untested version"  It must be restore the last known good state and evaluate how we allowed the broken version to merge.

Patch-set: 2
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0