Update patch set 4
Patch Set 4: Code-Review-1

Why were urllib and zhmclient held back (the 'why' is not in the commit message) (that's the reason for -1)

generate-constraints bot update includes the update to 2.20.1  The actual changes are VERY minor, only containing one commit with a code change.
https://github.com/requests/requests/compare/v2.20.0...v2.20.1
https://github.com/requests/requests/commit/ea9436a5d6a5934a906e91202349aabf9af75d15

I'd prefer to use the update in https://review.openstack.org/616780 since it's going to merge soon anyway (hopefully), but won't hold this one back.

Patch-set: 4
Label: Code-Review=-1