Update patch set 1
Patch Set 1:

The update bot uses it only for pushing requirements around - and that is not used anymore and that started this change.

I see no other users of projects.txt - and my open question is what to do with the integration job since it needs projects.txt - is it really testing what we need today? Tony, if you could look at that, it would be great!

Patch-set: 1