Update patch set 1
Patch Set 1: Code-Review-1

(1 comment)

This happens every time there is a new minor release of CentOS, and we typically say no as it's a much bigger change then expected.  Here's my understanding of the criteria:

 1) Tthat libvirt-python has to be >= libvirt.so
 2) libvirt-python listed in upper-constraints works with the libvirt.so in distros that consume this branch.  Including CentOS-7, Ubuntu LTS, SLES and highly desirable Gentoo.
 3) Validation that the new libvirt-python and libvirt.so are compatible with Nova's expectations.
 4) From the Extended Maintenance policy, the libvirt-python on branch n-1 be >= libvirt-python on branch n, and not spot backported.

So I'm -1 on this change until we can see the above is true.

Patch-set: 1
Label: Code-Review=-1