Update patch set 1
Patch Set 1: Code-Review-1

> > I'd consider it an integration test (of a specific subtype
 > > (upgrade))
 > 
 > OK, well it doesn't appear that the grenade repo has a grenade-py27
 > job or similar listed.  I would suggest you work with the OpenStack
 > QA team to help define something that meets your needs.
 > 
 > If you feel strongly about it, then I can transition the definition
 > of the job and neutron-grenade can become requirements-grenade-py27
 > if you wish.  You would then have ownership of it going forward.

Yup based on my reading of[1,2] we absolutely need grenade-py2 (which is functionally the same as neutron-grenade) to exist before and use that.  I'd also posit that *not* testing that on T would be questionable with reference to [1,2]

I get that the Neutron team wants to cease maintaining this job and template and I don't want to stop that but it's also not the neutron team's role to force this migration on all the existing consuming teams.

I suggest that the Neutron team, myself and the QA team work to creating grenade-py2 in the grenade repo and transition all existing users to that, and optionally add the py3 variant.

I've started this with: https://review.openstack.org/#/q/topic:grenade-py2  We need to do similar with neutron-grenade-multinode -> grenade-multinode-py2 and then migrate users to the new job names

[1] https://governance.openstack.org/tc/resolutions/20180529-python2-deprecation-timeline.html
[2] https://governance.openstack.org/tc/goals/stein/python3-first.html

Patch-set: 1
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1