Update patch set 1
Patch Set 1:

(1 comment)

Tony writes:
> This happens every time there is a new minor release of CentOS, and
we typically say no as it's a much bigger change then expected.

I think this is essentially the same situation that we discussed here:

https://review.openstack.org/#/c/578558/

Even though we can freeze the pip requirements for stable branches we can't freeze the OS and packages used to run the jobs in CI so eventually we have breakage (like we currently have with CentOS, but it could be any of the OS-es used in gate) and something has to be adjusted on the stable branches to make CI work again.

Patch-set: 1
Label: Workflow=0