Update patch set 2
Patch Set 2:

I'd also like to clarify my position.

I'm fine with accepting this change as long as it's done right.  From my POV we have a path forward we just need to get it done.  I kinda think that Tom is the lucky soul on the hook for this but I'm sure if LOCI need this he'd accept help :)

Patch-set: 2