Update patch set 3
Patch Set 3:

Here's the situation for SUSE: We're testing this in the gate against leap 15.0 and plan to update this to 15.1 when it becomes available.

Leap 15.0: libvirt 4.0.0
Leap 15.1: libvirt 4.10.0

So this patch doesn't make from a pure version perspective any difference. I wonder if therr is a real difference we're gonna hit. What exactly is failing right now? Which job exposes the issue?

Patch-set: 3
Label: Verified=0