Update patch set 4
Patch Set 4:

> What exactly is failing right now? Which job exposes the issue?

Dirk I raised a bug [1] pointing to a manila job that exposes the problem and maybe Chris can point to a loci job in that bug, but I think devstack will fail to install for every CentOS job on stable/rocky at the point where it tries to build a wheel for libvirt-python.

Next time I need to update this review I'll reference the bug in the commit message.

[1] https://bugs.launchpad.net/manila/+bug/1808342

Patch-set: 4