Update patch set 3
Patch Set 3:

> We have customers that would like the feature this enables to be available to them in Queens. Email to follow with more details.

That is not sufficient justification for breaking stable branch policy. We all have customers on older versions of openstack that would like newer features. It is the job of vendors, not the community, to figure out how to satisfy those requests for their customers, either by getting the customer upgraded or carrying internal forks on stable branches.

Patch-set: 3