Update patch set 3
Patch Set 3: Workflow-1

Yeah, we don't want to do this until we get a fixed version of grpcio. Sorry for not WIP'ing it.

I guess my only concern with an explicit list is that if they do a bugfix release on one of those branches then there's a new version that won't work, but that may not be a big problem. Anyone installing from pip will get a newer version, and presumably the distros would have avoided the broken versions in the first place if they were following our g-r list. You'd have to make a pretty serious effort to get to a broken bugfix release.

Patch-set: 3
Label: Workflow=-1