Add Yappi to global-requirements
Yappi is needed to profile (capture function calltrace) oslo.service
processes.

Is the library actively maintained?
Yes
Is the library good code?
Yes
Is the library python 3 compatible?
Yes, it has py3 gates
Is the library license compatible?
Yes
Is the function of this library already covered by other libraries in global-requirements.txt?
No
Is the library required for OpenStack project or related dev or infrastructure setup? (Answer to this should be Yes, of course) Which?
Yes, oslo.service
https://review.openstack.org/#/c/627414/

Change-Id: Ibea0cdb732923f1b53d5cb6aeeb4041fb5973494