Update patch set 1
Patch Set 1:

Matthew, there is a genuine reason for this: thre are projects that do not have per-release branches but they do use openstack requirements based on the tested target release.

This means that we need to backport any new blacklist in order to be able to run their code.

Patch-set: 1