Update patch set 1
Patch Set 1: Code-Review+1

As I noted in the bug, it doesn't appear that this is strictly a bug in privsep, but more a case of the threading change exposing some code that didn't play nicely with threads. Unfortunately, as this completely breaks the relevant Neutron module I think we need to go ahead with this for now.

However, capping it also blocks Cinder from further adopting privsep so we need to find a proper fix ASAP.

Patch-set: 1
Reviewer: Gerrit User 6928 <6928@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+1