Update patch set 1
Patch Set 1:

I am not sure I understand that question. Still, I do not see any reason for not adding the missing linter to the blacklist.

The reason is simple: if linting changes are not backported it will be really hard to backport other changes, mainly because we would likely
lose the ability to "cherrypick to branch" due to conflicts.

AFAIK, blacklist file contains linters so I see no restrictions for updating their list. Nothing got shipped from here, but they are needed for testing the code, on all supported branches.

Patch-set: 1