Update patch set 1
Patch Set 1: Code-Review+2

given it's linter and it's blacklist we should be fine adding it, but as sean asked, we do need to figure out what to do with 'non-stable-follows-policy' projects using requirements (which is for stable-follows-policy as one of it's objectives).  I'd like the second vote here to be from the stable team (or at least an ack).  I'll bring this up at our meeting today.

Patch-set: 1
Reviewer: Gerrit User 14288 <14288@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=+2