Update patch set 1
Abandoned

Thanks Matthew for the information!

Then I guess *for stable branches* now the best solution is to restrict neutron on the given project instead of here. (As the replacement of the neutron to neutron-lib is rather a new feature, that fits to master branches, i guess)

Patch-set: 1
Status: abandoned