Update patch set 1
Patch Set 1: Code-Review-1

Is sahara a service project (is it cycle with intermediary)?

The way we recommend code needed for plugins to be handled is to make a foo-lib package, that is then depended upon by those that need it (plugins and/or service).

We also need the answer to the following linked questions
https://docs.openstack.org/project-team-guide/dependency-management.html#for-new-requirements

Patch-set: 1
Label: Code-Review=-1