Update patch set 1
Patch Set 1:

I'd personally like to avoid another split for time reasons, but this is a decision of the PTL (in CC in the review).

I would also argue that the problems raised by other services should not apply to the way the sahara split was organized: we don't plan internal changes to the plugin interface.

In any case we need to have a way for plugins to depend on sahara before the release of Stein.

Patch-set: 1