Update patch set 1
Patch Set 1:

I'm pretty sure that this will actually break ppc64le as https://github.com/python-greenlet/greenlet/commit/b129330662c76cab62f6f3e718071834a8546ee1 is wrong.

We kinda need the owners of the ppc64le toolchain to weigh in on that.  One way to encourage that would be to merge this and then wait to see what breaks ?

#stumped

Patch-set: 1