Update patch set 1
Patch Set 1: Code-Review-1

-SQLAlchemy>=0.7.8,!=0.9.5
+SQLAlchemy>=0.9.6
This one may hurt projects, we'd be changing the min supported version of sqlalchemy

-    setup_requires=['pbr>=1.3'],
+    setup_requires=['pbr>=1.8'],

-install_command = pip install {opts} {packages}
+# Avoid psycopg2 wheel package rename warnings by not using the binary.
+install_command = pip install --no-binary psycopg2 {opts} {packages}
https://github.com/openstack/sqlalchemy-migrate/commit/f44d07956a2827484cddeee161cbd7724229f92f
doesn't seem like a big deal will build from source if needed

py26 support was removed, but we don't care I don't think

I think one of the following 3 was the fix, can you let me know which?  The last one probably can't be backported (as it changes deps) but the other two could be for a 0.11.1 release upstream and be consumed by us at that point.

https://github.com/openstack/sqlalchemy-migrate/commit/a00dab7bcfb815ce370e14d1b33f328d5aee1c0d
https://github.com/openstack/sqlalchemy-migrate/commit/231a4d2ae9f8496cfc8eea2e6bb8186a0dd602f9
https://github.com/openstack/sqlalchemy-migrate/commit/fe646671060815ff865e49b52d957c7f01220597

https://github.com/openstack/sqlalchemy-migrate/compare/0.11.0...0.12.0

Patch-set: 1