Update patch set 1
Patch Set 1:

> adding this here so we know it should be possible to make a 0.11.1
 > 
 > git checkout 0.11.0
 > git cherry-pick 231a4d2ae9f8496cfc8eea2e6bb8186a0dd602f9
 > make other 0.11.1 change
 > git commit
 > git tag

It's slightly more complex / different to that and given that we haven;t tested releasing from stable branches I'd expect there to be a problems.

Having said that stable branches could be a worthwhile and I'm happy to work with the sqlalchemy-migrate cores and infra to make that happen.

Patch-set: 1