Update patch set 2
Patch Set 2: Code-Review-1

ok, I looked a litte and the last update by the main dev was august 2016. coreycb commented on it too, as a side note.

The lack of response since then is why I'm voting against this, if there is some indication that they are still active I'll approve.

Patch-set: 2