Update patch set 2
Patch Set 2:

I’m also a bit worried about the inactivity. Two things swayed me to submit this anyway:  1) I couldn’t find any well used versions that were active, and 2) it’s already in upper constraints, due to being pulled in by some other dependency, so we are already on the hook for it. Also, the distros actively package it.

Patch-set: 2