Update patch set 2
Patch Set 2: Code-Review-2

We discussed this on the list, starting here:

http://lists.openstack.org/pipermail/openstack-discuss/2019-February/002772.html

in that discussion we came to the conclusion that relying on the requirements team for security updates was not a good idea.  That, along with the fact that many times updating requires large updates (as it does here).  If you'd like to change that, please post in that thread.

Patch-set: 2