Update patch set 1
Patch Set 1:

At first glance, they are not using their requirements.txt for their actual requirements, which is good (as they pin and we don't support staticly pinned libs).  'requests>=2.21.0' and possibly 'parser': ['pyhcl>=0.3.10'].   If the parser is pulled in it WILL require a pin on ply, which is no good ( https://github.com/virtuald/pyhcl/blob/master/requirements.txt ).  It'd pass testing (same as what we have, for now), but then it'd lock all of openstack to that version for UC.txt.

Patch-set: 1