Fix build all upper-constraints wheels w/ py3.6
zVMCloudConnector wheel cannot be built on python3.6
with a version below 1.0.0. Yet, the global-requirements
would allow to try building wheels for lower versions
than 1.0.0.

This is a problem, as LOCI project builds wheels for all the
packages in upper-constraints, and then consumes those when
installing projects. For those with python3.6, this would cause
a failure to build wheels, and an inability to consume those
wheels.

The problem does not appear on Rocky, as the
minimum version has been bumped there for python3.6 already.

This fixes it by conditionally bumping the version for
zVMCloudConnector on the python version.

Notes:
- Projects consuming zVMCloudConnector in Queens probably
  will not be interested in bumping their version of
  zVMCloudConnector for python3.6, therefore the fix should come
  from consuming projects/requirements.
- Bumping this version merely allows buildability of the wheels,
  not that the project will work with python3.6. For that,
  the necessary patches have to be applied, and a bigger bump is
  required. However, I consider running software for python3.6 is
  a completely different project in itself than this patch.

Change-Id: Iae4ec550de101648ff9c8d1b743ccaa285bd0033