Update patch set 1
Patch Set 1: Code-Review-1

This impacts multiple vendors in ways we said we wouldn't do so we need to look harder.

Well isn't this a fun little problem :D

My thoughts on this are going to wander all over the place and frankly they're not at all well formed so I expect they'll be hard to follow ... sorry.

My first thought is openstack/requirements is about ensuring that OpenStack components are co-installable and at least one[1] of them is pinned to 2.4.  I'm pretty sure that this is about updating the ansible used buy the gate but we'd like very much for that to be compatible with any ansible needed by OpenStack projects?

[2] Says that 2.5.9 is the current 2.5 release and is security supported .. Oh I see when 2.8 is released 2.5 goes EOL and I assume given ~4 month release cycle 2.8.0 will be ready Real Soon Now.

Given our new(ish) Extended Maintenance cycle queens could and probably will be a thing in 4 months so will we need to move to 2.7 when that happens?

IIUC the 0.12.0 requirement comes from [3] so that'd mean that os_quota[4] isn't and can't work in an ansible installed with u-c.

I think we need to understand most? of that ... Assuming that we do and we all agree we need something like this

The biggest issue is that 0.12.0 is the first rocky[4] release and introduces the following effective minimum bumps:

 [tony@thor openstacksdk]$ git diff -U0 0.11.0..0.12.0 -- *requirements.txt 
 diff --git a/requirements.txt b/requirements.txt
 index f131da00..7a4f9fed 100644
 --- a/requirements.txt
 +++ b/requirements.txt
 @@ -5 +5 @@ pbr!=2.1.0,>=2.0.0 # Apache-2.0
 -PyYAML>=3.10 # MIT
 +PyYAML>=3.12 # MIT
 @@ -10,2 +10,2 @@ six>=1.10.0 # MIT
 -os-service-types>=1.1.0 # Apache-2.0
 -keystoneauth1>=3.3.0 # Apache-2.0
 +os-service-types>=1.2.0 # Apache-2.0
 +keystoneauth1>=3.4.0 # Apache-2.0
 @@ -17 +17 @@ jmespath>=0.9.0 # MIT
 -ipaddress>=1.0.16;python_version<'3.3' # PSF
 +ipaddress>=1.0.17;python_version<'3.3' # PSF

The only one of which that isn't covered by the queens u-c is os-service-types which implies a bump to >= 1.2.0

 [tony@thor os-service-types]$ git diff -U0 1.1.0..1.2.0 -- *requirements.txt 
 diff --git a/test-requirements.txt b/test-requirements.txt
 index e3f6dd2..a94411c 100644
 --- a/test-requirements.txt
 +++ b/test-requirements.txt
 @@ -8,3 +8,3 @@ coverage!=4.4,>=4.0 # Apache-2.0
 -python-subunit>=0.0.18 # Apache-2.0/BSD
 -sphinx>=1.6.2 # BSD
 -oslotest>=1.10.0 # Apache-2.0
 +python-subunit>=1.0.0 # Apache-2.0/BSD
 +sphinx!=1.6.6,>=1.6.2 # BSD
 +oslotest>=3.2.0 # Apache-2.0
 @@ -12,3 +12,3 @@ testscenarios>=0.4 # Apache-2.0/BSD
 -requests-mock>=1.1 # Apache-2.0
 -openstackdocstheme>=1.16.0 # Apache-2.0
 -keystoneauth1>=3.1.0 # Apache-2.0
 +requests-mock>=1.1.0 # Apache-2.0
 +openstackdocstheme>=1.18.1 # Apache-2.0
 +keystoneauth1>=3.3.0 # Apache-2.0
 @@ -16 +16 @@ keystoneauth1>=3.1.0 # Apache-2.0
 -reno!=2.3.1,>=1.8.0 # Apache-2.0
 +reno>=2.5.0 # Apache-2.0

which are test-only changes and all of which seem to be from queens anyway?

Coming back to the effective bumps.  We're getting better with lower-constraints at covering and testing them but we know we're not there for rocky (and probably still not for stein)

So if we were to do this we would be asking vendors to bump the shipped version.

 Debian: seems to be 1.1.0 [6]
 RDO: is 1.1.0 [7]
 Gentoo: seems to have dropped queens? [8]
 suse:  I don't know how to map [9] to released things :/ sorry
 ubuntu: is 1.1.0 [10]

I haven't looked at the 'enterprise' releases like suse cloud or RedHat OpenStack platform but I'd expect them to be similar to the community releases 

So that's the long way round of saying not without more information[11]

[1] https://opendev.org/openstack/openstack-ansible/src/branch/stable/queens/scripts/bootstrap-ansible.sh#L25
[2] https://docs.ansible.com/ansible/latest/reference_appendices/release_and_maintenance.html
[3] https://github.com/ansible/ansible/blob/stable-2.6/lib/ansible/module_utils/openstack.py#L111
[4] https://github.com/ansible/ansible/blob/stable-2.6/lib/ansible/modules/cloud/openstack/os_quota.py#L109
[5] https://opendev.org/openstack/releases/src/branch/master/deliverables/rocky/openstacksdk.yaml#L9-L12
[6] https://salsa.debian.org/openstack-team/libs/python-os-service-types/commit/f130e94dbed08d6dc9185dae00b109e86b4b8bce
[7] https://trunk.rdoproject.org/centos7-queens/current/python-os-service-types-1.1.0-0.20180211211248.f130e94.el7.centos.src.rpm
[8] https://packages.gentoo.org/packages/dev-python/os-service-types
[9] https://build.opensuse.org/package/show/Cloud:OpenStack:Upstream:Master/python-os-service-types
[10] https://launchpad.net/~ubuntu-cloud-archive/+archive/ubuntu/queens-staging/+index?batch=75&memo=150&start=150
[11] I think this is the most verbose review comment I've ever made

Patch-set: 1
Reviewer: Gerrit User 12898 <12898@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1