networking-ansible has broken out api code to its own library: network-runner
This patch adds network-runner.
Networking-ansible is in process of replacing the ansible-runner library
with the network-runner library.

We have decided to keep the existing api code that depends on ansible-runner
through the train cycle. During this cycle we will require on both libraries.
In the U release we will remove the old api that depends on ansible-runner.

Ansible-runner is also now required by flame and windmill-ops.

Is the library actively maintained?
Yes. It's maintained by the same folks maintaining networking-ansible

Is the library good code?
We think so.

Is the library python 3 compatible?
Yes

Is the library license compatible?
ASL 2.0

Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?
Fedora Yes
Ubuntu No

Is the function of this library already covered by other libraries in global-requirements.txt?
Nope

Is the library required for OpenStack project or related dev or infrastructure setup? (Answer to this should be Yes, of course) Which?
Yes, networking-ansible requires this package.

If the library release is managed by the Openstack release process does it use the cycle-with-intermediary release type?
Not managed by Openstack release cycle

Do I need to update anything else?
Don't know of anything

Change-Id: I5ecb9fa79176c65fc9a9460a5cd0059508bd2701