networking-ansible has broken out api code to its own library: network-runner
This patch removes the ansible-runner dependency that
networking-ansible was using and replaces it with network-runner.

Is the library actively maintained?
Yes. It's maintained by the same folks maintaining networking-ansible

Is the library good code?
We think so.

Is the library python 3 compatible?
Yes

Is the library license compatible?
ASL 2.0

Is the library already packaged in the distros we target (Ubuntu latest / Fedora latest)?
Fedora Yes
Ubuntu No

Is the function of this library already covered by other libraries in global-requirements.txt?
Nope

Is the library required for OpenStack project or related dev or infrastructure setup? (Answer to this should be Yes, of course) Which?
Yes, networking-ansible requires this package.

If the library release is managed by the Openstack release process does it use the cycle-with-intermediary release type?
Not managed by Openstack release cycle

Do I need to update anything else?
Don't know of anything

Change-Id: I5ecb9fa79176c65fc9a9460a5cd0059508bd2701