Update patch set 1
Patch Set 1:

vmware-nsx appears to be using octavia code directly and accessing the octavia database. Both of these are bad practice and should not be done. octavia was never in global-requirements and should not be.
This code needs to be fixed to only use the octavia-lib and/or the octavia APIs.

Patch-set: 1