Update patch set 2
Patch Set 2:

not too much code and in browsing it it semed ok.  The unit tests may be a little light, not sure about the *.py_ files in there, haven't seen that before.
https://github.com/ansible-network/network-runner/tree/devel/network_runner/tests/unit

It looks like this depends on ansible itself as well (looking at both requirements.txt and test-requirements.txt) I don't see that in our upper-constraints.  I'm not sure what it'd pull in as well, if ansible itself pulls in it's extra_requires stuff it could be a ton of new libraries.  I'm not sure, but is ansible supposed to be used as a library or do they have that split out to some helper that is better exportable?

tox itself seems fine

Patch-set: 2
Label: Verified=0