Update patch set 3
Patch Set 3:

This has been in the check queue for almost a week now. Apparently we lost our only arm cloud resources and are waiting for a new provider to be set up. Not sure if we want to abandon this one for now, or just let it sit out there and wait for those new nodes to become available.

Patch-set: 3
Reviewer: Gerrit User 11904 <11904@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0