Update patch set 1
Patch Set 1:

Actually, seems it's not harmful for the upstream CI. It just influences our company's internal CI. We use branch to track the project and use upstream requirements to limit the version. But we found our internal stable/pike don't have 2.32.2 tag. So it will cause failure. So if this can be fixed on community wii be better. Of course, if we think we don't need to do this revert. Feel free to abandon this patch.

regards,
Damon Li

Patch-set: 1
Reviewer: Gerrit User 22078 <22078@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0