Update patch set 1
Patch Set 1:

Typically when we add a new requirement, we want a set of questions answered about each one:

https://docs.openstack.org/project-team-guide/dependency-management.html#for-new-requirements

This is a little different though. These really have been dependencies of Cinder for a long time now, we just handled the management of them poorly and this is now attempting to get things right.

I'll leave it up to Matt if we need to break this out for each and have those questions answered.

Patch-set: 1
Reviewer: Gerrit User 11904 <11904@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0