Update patch set 2
Patch Set 2:

I think we still need the questions answered, at the very least we need the licences checked, but if we are adding it to GR that kinda means we looked at it at least once for maintainability, testing, activity, etc.  It'd probably be easiest to do this all at once, the licence info is in the commit (GR comments), so that's one step.

Patch-set: 2