Update patch set 2
Patch Set 2:

I don't know the answers to the questions as I don't maintain the libraries.   I can vouch for the hpe python libs as I created those years ago, but without this patch, we cannot add the optional extras to cinder to make sure the required libraries are available for our drivers.

Patch-set: 2
Label: Verified=0