Cap keyring and sphinx on python2
keyring drops py2 support so lets ensure our requirements wont allow us
to install 19.X releases on py2.

Setting the minimum supported python for both to 3.4 to be consistent
with u-c.  When we remove the 3.4 constraints we'll also need to fix g-r

Similarly sphinux 2.0.0 drop py27 support do express that in
global-requirements.

There isn't a need to update u-c as they're already correct:
  $ grep -E '^(Sphinx|keyring)' upper-constraints.txt
  keyring===18.0.1;python_version=='2.7'
  keyring===19.0.1;python_version=='3.4'
  keyring===19.0.1;python_version=='3.5'
  keyring===19.0.1;python_version=='3.6'
  Sphinx===1.8.5;python_version=='2.7'
  Sphinx===2.0.1;python_version=='3.4'
  Sphinx===2.0.1;python_version=='3.5'
  Sphinx===2.0.1;python_version=='3.6'

This may cause knock-on effects for projects using either of these in
lower-constraints, but that's less impactful than all the 'red' in the
gate right now

Also 'correct' the Sphinx requirement in test-requirements so we get the
correct version when running under python2.  We need to do this anyway as
docs jobs typically run unconstrained

Change-Id: If558f184c959e4b63b56dec3ca1571d1034cfe5c