Update patch set 1
Patch Set 1:

half developed?  That library has been around for years and used by the 3par driver in cinder for a very long time.  It's not 'half developed'.  It's a fully featured client library for the 3PAR storage devices that I originally authored back in 2012.  

Also, fwiw, I added the py37 to tox.ini and ran the tests and they passed just fine.

http://paste.openstack.org/show/5TPPZ7E0yIj1K4iyLCEZ/

I have a PR in to add py36 and py37 to tox.ini here:
https://github.com/hpe-storage/python-3parclient/pull/65

Patch-set: 1