Update patch set 1
Patch Set 1:

> Wouldn't it be better to update the generate target to override its
 > install_command?
 > 
 > I see a wide mix of this in other projects with the constraints
 > being in either install_command or deps. I suppose it is only an
 > issue for us though since we are probably the only ones that need
 > certain jobs to run unconstrained.

Using constraints in install_command is wrong and will break any project that is also listed in constraints.

For our use case adding it into install_command for the generate testenv will work but IMO introduces a small amount of cognitive dissidence so I prefer to keep all projects following the same "template".

Hopefully once the constraints-update topic is empty on master we wont see that anymore.  (If we do we should fix it)

Patch-set: 1