Update patch set 1
Patch Set 1:

> Thanks Matthew and Tony for the reviews. I don't have any issue
 > with this. It's just that when trying to fix sphinx issue in
 > Searchlight [1] by selecting python version, I think It'd better
 > now use py3.6 and above so we don't have to change it later.
 > 
 > [1] https://review.opendev.org/#/c/659501/

Oh :(

So cards on the table, what you want to do is the right thing.  However we (the requirements team) are't ready to make the changes needed to remove py<3.6.  So we're kinda at an impasse .... 659501 can't merge yet and we can't merge this yet :(

With all the constraints work on our plate ATM I doubt we'll get to the work you need anytime soon, I'm guessing it'll be around m2

Are you willing to update 659501 to say <3.4, and then add a patch on top that says <3.6 (yes that's fixing it later but with the review there it wont be forgotten)

Patch-set: 1