Update patch set 2
Patch Set 2:

There are two options, in order of prefrence...

1 make nova-lib for exportable classes/functions
1.1 used by nova and not-nova
1.2 cycle with intermediary release type

2 switch nova to cycle with intermediary

The reasoning for this is that if you use something as a library you have to treat it as one.  See what octavia is doing (making octavia-lib), we've gotten reviews to add octavia to global-reqs that have been shot down for similiar reasons (by reqs team and octavia ptl).

Patch-set: 2