Update patch set 3
Patch Set 3:

> (1 comment)

As for the questions, I'm not sure much if anything applies. tripleo-ansible is a collection of roles and playbooks and while it can be packaged for redistribution via python, that is not it's primary purpose. Also we're not providing any python library or depending on any new libraries, we're just wanting requirements to maintain our requirements.txt and test-requirements.txt files.

Patch-set: 3
Reviewer: Gerrit User 7353 <7353@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0