Add scapy to global requirements
In a new feature in Neutron, a user will be able to add ethertypes other
than just IPv4 and IPv6 as part of security groups.  In order to
properly test this, a fullstack test is being created that sets up a
tcpdump and then crafts a packet with a custom ethertype, to make sure
that the packet is detected.  The library we use to create the packet
with the custom ethertype, scapy, needs to be added to the global
requirements list.

New requirement questionnaire:

1.) Is the library actively maintained?

Yes, the library is under active development as you can see from the
commit frequency graph [1].

2.) Is the library good code?

Yes.  Scapy is commonly used in the security community to craft packets
(the same use we would like for it here) and is considered the best
library for this purpose in Python from those I have talked to.

3.) Is the library python 3 compatible?

Yes, Scapy supports python 2.7 as well as python 3.4 through 3.7.

4.) Is the library license compatible?

Scapy is licensed under the GPL v2, which is listed as an incompatible
license per [2].  Since this is a testing use case for this library, the
instructions indicate that this should be sent to the legal-discuss
list.

The primary alternative for this library is PyPacker [3] which is also
GPL v2.

5.) Is the library already packaged in the distros we target (Ubuntu
latest / Fedora latest)?

Yes it is [4].

6.) Is the function of this library already covered by other libraries
in global-requirements.txt?

I have not reviewed all 500 or so libraries in global-requirements.txt,
but no I am not aware of any other library included in
global-requirements.txt that allows for construction of arbitrary
packets.  If there is one I would be happy to use it.

7.) Is the library required for OpenStack project or related dev or
infrastructure setup? Which?

Yes, this is required for OpenStack Networking (neutron) to provide
fullstack testing for the custom ethertypes feature.  The custom
ethertypes feature spec [5] has a full description.  For more
information see the top paragraph here.

8.) If the library release is managed by the Openstack release process
does it use the cycle-with-intermediary release type?

This library is released outside the OpenStack release process.

9.) Do I need to update anything else?

Projects needing to use scapy will need to add tcpdump to their
bindep.txt.

[1] https://github.com/secdev/scapy/graphs/commit-activity
[2] https://governance.openstack.org/tc/reference/licensing.html
[3] https://gitlab.com/mike01/pypacker
[4] https://repology.org/project/scapy/versions
[5] https://specs.openstack.org/openstack/neutron-specs/specs/train/custom_ethertypes.html

Change-Id: I1b8cde55e36ca476e6f64b3f414151ceb40feeb0
Related-Bug: #1832758
Needed-By: https://review.opendev.org/670203