Add scapy to global requirements
In a new feature in Neutron, a user will be able to add ethertypes other
than just IPv4 and IPv6 as part of security groups.  In order to
properly test this, a fullstack test is being created that sets up a
tcpdump and then crafts a packet with a custom ethertype, to make sure
that the packet is detected.  The library we use to create the packet
with the custom ethertype, scapy, needs to be added to the global
requirements list.

Change-Id: I1b8cde55e36ca476e6f64b3f414151ceb40feeb0
Related-Bug: #1832758
Needed-By: https://review.opendev.org/670203