Update patch set 1
Patch Set 1: Code-Review-1

do yo mind adding it to upper-constraints as well?  It looks like it has optional dependencies, not sure if we pull those in on constraints updates, but it may be odd to track ipython.  The main concern would by pulling in pyx, which while we don't track it as a first level dependency (global-requirements.txt) it would be brought in as a constraint (it's GPL).

We also need https://docs.openstack.org/project-team-guide/dependency-management.html#for-new-requirements to be answered in commit message, thanks (a link to the spec could be nice)..

Patch-set: 1
Reviewer: Gerrit User 14288 <14288@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Code-Review=-1