Update patch set 9
Patch Set 9:

It looks like the grenade job is failing because of a testtools issue: https://github.com/testing-cabal/testtools/issues/272 (well the reverse case of what I documented there unittest run with a testtools suite vs testtools run with a unittest suite). The stestr release added here, 2.5.0[1], switched the internal test runner it was using to be stdlib unittest based for python 3 instead of testtools based. This was done to fix a myriad of bugs for people who don't use testtools (which outside of openstack is no widely used) and was a constant source of bugs and complaints from non-openstack stestr users.

I'm not sure what the best path forward to fix that bug is. I can't really revert the stestr internal test runner change because it was necessary to fix critical issues for other projects. Testtools is in a pretty stale state because of a lack of maintenance (I was recently given maintainer privileges on it, but lack the time to really do anything with it) it's unlikely to be fixed over there anytime soon. I do have an ideal long term fix proposed here: https://github.com/testing-cabal/testtools/pull/277 but that can't merge until python2  and python 3.4 support is dropped from testtools. Maybe we can come up with a smaller more self contained fix to workaround unittest2 not playing well with stdlib unittest in newer python 3 and get that merged in testtools and I can push out a quick maint. release.

[1] https://github.com/mtreinish/stestr/releases/tag/2.5.0

Patch-set: 9
Reviewer: Gerrit User 5196 <5196@4a232e18-c5a9-48ee-94c0-e04e7cca6543>
Label: Verified=0